Moving typealias from sample to NumpySample
