+++ ocf_data_sampler/numpy_sample/base.py
