+++ ocf_data_sampler/numpy_sample/nwp.py
@@ -3,7 +3,7 @@
 import pandas as pd
 import xarray as xr
 
-from ocf_data_sampler.sample.base import NumpySample
+from ocf_data_sampler.numpy_sample.base import NumpySample
 
 
 class NWPSampleKey:
