+++ ocf_data_sampler/sample/__init__.py
@@ -1,3 +1,3 @@
-from ocf_data_sampler.sample.base import SampleBase
+from ocf_data_sampler.numpy_sample.base import SampleBase
 from ocf_data_sampler.sample.uk_regional import UKRegionalSample
 from ocf_data_sampler.sample.site import SiteSample
